Factorial Digit Sum
https://www.hackerrank.com/contests/projecteuler/challenges/euler020

This was stunningly easy for something with <80% pass rate on hackerrank.  Maybe something about this being easier in Python 2.
In addition to the builti-in factorial, guess the ability to easily break a digit into numbers and sum them helped.
------
n! means n×(n−1)×⋯×3×2×1
For example, 10!=10×9×⋯×3×2×1=3628800, 
and the sum of the digits in the number 10! is 3+6+2+8+8+0+0=27.

Find the sum of the digits in the number N!
Input Format 
The first line contains an integer T , i.e., number of test cases. 
Next T lines will contain an integer N.

Output Format 
Print the values corresponding to each test case.

Constraints 
1≤T≤100 
0≤N≤1000

Sample Input
2
3
6
Sample Output
6
9