Coded Triangular Numbers
https://www.hackerrank.com/contests/projecteuler/challenges/euler042/

This was really more Math than Python
Good to review quadratic formula from High School though
------------------------------------------------------------
The nth term of a sequence of triangle numbers is given by, 
tn=n^2+n/ 2

so the first ten triangle numbers are: 
1,3,6,10,15,21,28,36,45,55,⋯
You are given an integer. If it is a triangular number tn, print the term n corresponding to this number, else print −1

Input Format 
First line of input contains an integer T denoting the number of testcases. Each of the next T lines contains an integer.

Output Format 
Print the answer corresponding to each test case in a new line.

Constraints 
1≤T≤105 
1≤tn≤1018

Sample Input #00:
3
2
3
55

Sample Output #00:
-1
2
10