Highly Divisible Triangular Number
https://www.hackerrank.com/contests/projecteuler/challenges/euler012/
-------------------------------
The math was interesting in this one (tried brute-forcing it first, which times out...before discovering the property that the # of factors
would be the product of # factors for N and # factors for (N+1)/2  when N is ODD, or 
# factors for N/2  X # factors for N+1  when N is EVEN

the OrderedDict wasn't technically needed.  Kept in the code because there were a few pythonic things
* the OrderedDict to sort a dictionary
* the min syntax
* exception handling
...that were useful/point of doing all these Python exercises.
-------------------------------

The sequence of triangle numbers is generated by adding the natural numbers. So the 7'th triangle number would be 1+2+3+4+5+6+7=28. The first ten terms would be:

1,3,6,10,15,21,28,36,45,55,...

Let us list the factors of the first seven triangle numbers:

1:1
3:1,3
6:1,2,3,6
10:1,2,5,10 
15:1,3,5,15 
21:1,3,7,21 
28:1,2,4,7,14,28

We can see that 28 is the first triangle number to have over five divisors.

What is the value of the first triangle number to have over N divisors?

Input 
First line TT, the number of testcases. Each testcase consists of NN in one line.

Output 
For each testcase, print the required answer in one line.

Constraints 
1≤T≤101≤T≤10 
1≤N≤1031≤N≤103

Sample input
4
1
2
3
4

Sample output
3
6
6
28