Reciprocal Cycles (Partial Solution, fails test case 4)
https://www.hackerrank.com/contests/projecteuler/challenges/euler026

This needs to be rewritten to use Sieve of Eratosthenes
The reason it's failing is because the precision (10000) is only sufficient to cover up to 5000.
But the problem set could pass in #s as high as 10000.  If I set the precision to 20000, it will time out > 10 seconds
See also https://www.engert.us/erwin/miscellaneous/Prime%20reciprocals%20repeating%20digits.pdf
------------------------------------------------

A unit fraction contains 1 in the numerator. The decimal representation of the unit fractions with denominators 2 to 10 are given:

1/2 = 0.5
1/3 = 0.(3)
1/4 = 0.25
1/5 = 0.2
1/6 = 0.1(6)
1/7 = 0.(142857)
1/8 = 0.125
1/9 = 0.(1)
1/10 = 0.1

Where 0.1(6) means 0.166666..., and has a 1-digit recurring cycle. It can be seen that 1/7 has a 6-digit recurring cycle.

Find the value of smallest d<N for which 1/d contains the longest recurring cycle in its decimal fraction part.

Input Format 
The first line contains an integer T , i.e., number of test cases. 
Next T lines will contain an integer N.

Output Format 
Print the values corresponding to each test case.

Constraints 
1≤T≤1000
4≤N≤10000


Sample Input
2
5
10

Sample Output
3
7